+++ internal/component/prometheus/fanout.go
@@ -32,6 +32,7 @@ type Fanout struct {
 	writeLatency   prometheus.Histogram
 	samplesCounter prometheus.Counter
 	ls             labelstore.LabelStore
+	metadataStore  UpdateableMetadataStore
 
 	// lastSeriesCount stores the number of series that were sent through the last appender. It helps to estimate how
 	// much memory to allocate for the staleness trackers.
@@ -39,7 +40,7 @@ type Fanout struct {
 }
 
 // NewFanout creates a fanout appendable.
-func NewFanout(children []storage.Appendable, componentID string, register prometheus.Registerer, ls labelstore.LabelStore) *Fanout {
+func NewFanout(children []storage.Appendable, componentID string, register prometheus.Registerer, ls labelstore.LabelStore, ms UpdateableMetadataStore) *Fanout {
 	wl := prometheus.NewHistogram(prometheus.HistogramOpts{
 		Name:    "prometheus_fanout_latency",
 		Help:    "Write latency for sending to direct and indirect components",
@@ -59,6 +60,7 @@ func NewFanout(children []storage.Appendable, componentID string, register prome
 		writeLatency:   wl,
 		samplesCounter: s,
 		ls:             ls,
+		metadataStore:  ms,
 	}
 }
 
@@ -79,13 +81,15 @@ func (f *Fanout) Appender(ctx context.Context) storage.Appender {
 	// contain both a scrape target and a metadata store, and fails the
 	// conversion if they are missing. We should find a way around this as both
 	// Targets and Metadata will be handled in a different way in Alloy.
+	// TODO(@ptodev): Can we instead use the more recent translator package:
+	// https://github.com/prometheus/otlptranslator
 	ctx = scrape.ContextWithTarget(ctx, scrape.NewTarget(
 		labels.EmptyLabels(),
 		&config.DefaultScrapeConfig,
 		model.LabelSet{},
 		model.LabelSet{},
 	))
-	ctx = scrape.ContextWithMetricMetadataStore(ctx, NoopMetadataStore{})
+	ctx = scrape.ContextWithMetricMetadataStore(ctx, f.metadataStore)
 
 	app := &appender{
 		children:          make([]storage.Appender, 0),
@@ -210,6 +214,13 @@ func (a *appender) UpdateMetadata(ref storage.SeriesRef, l labels.Labels, m meta
 	if ref == 0 {
 		ref = storage.SeriesRef(a.fanout.ls.GetOrAddGlobalRefID(l))
 	}
+
+	// Store metadata in our local store
+	familyName := l.Get(model.MetricNameLabel)
+	if familyName != "" {
+		a.fanout.metadataStore.UpdateMetadata(familyName, m)
+	}
+
 	var multiErr error
 	for _, x := range a.children {
 		_, err := x.UpdateMetadata(ref, l, m)
@@ -270,20 +281,3 @@ func (a *appender) AppendHistogramCTZeroSample(ref storage.SeriesRef, l labels.L
 	}
 	return ref, multiErr
 }
-
-// NoopMetadataStore implements the MetricMetadataStore interface.
-type NoopMetadataStore map[string]scrape.MetricMetadata
-
-// GetMetadata implements the MetricMetadataStore interface.
-func (ms NoopMetadataStore) GetMetadata(_ string) (scrape.MetricMetadata, bool) {
-	return scrape.MetricMetadata{}, false
-}
-
-// ListMetadata implements the MetricMetadataStore interface.
-func (ms NoopMetadataStore) ListMetadata() []scrape.MetricMetadata { return nil }
-
-// SizeMetadata implements the MetricMetadataStore interface.
-func (ms NoopMetadataStore) SizeMetadata() int { return 0 }
-
-// LengthMetadata implements the MetricMetadataStore interface.
-func (ms NoopMetadataStore) LengthMetadata() int { return 0 }